Shift Contact related parser classes into contact list parser package
