--- src/main/java/seedu/address/logic/parser/DeleteContactParser.java
@@ -1,28 +1,0 @@
-package seedu.address.logic.parser;
-
-import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-
-import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.contactlistcommands.DeleteContactCommand;
-import seedu.address.logic.parser.exceptions.ParseException;
-
-/**
- * Parses input arguments and creates a new DeleteContactCommand object
- */
-public class DeleteContactParser implements Parser<DeleteContactCommand> {
-
-    /**
-     * Parses the given {@code String} of arguments in the context of the DeleteContactCommand
-     * and returns a DeleteContactCommand object for execution.
-     * @throws ParseException if the user input does not conform the expected format
-     */
-    public DeleteContactCommand parse(String args) throws ParseException {
-        try {
-            Index index = ParserUtil.parseIndex(args);
-            return new DeleteContactCommand(index);
-        } catch (ParseException pe) {
-            throw new ParseException(
-                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteContactCommand.MESSAGE_USAGE), pe);
-        }
-    }
-}
