--- src/main/java/seedu/address/logic/parser/EditContactParser.java
@@ -1,58 +1,0 @@
-package seedu.address.logic.parser;
-
-import static java.util.Objects.requireNonNull;
-import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_TELEGRAM;
-
-import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.contactlistcommands.EditContactCommand;
-import seedu.address.logic.commands.contactlistcommands.EditContactDescriptor;
-import seedu.address.logic.parser.exceptions.ParseException;
-
-/**
- * Parses input arguments and creates a new EditContactCommand object
- */
-public class EditContactParser implements Parser<EditContactCommand> {
-
-    /**
-     * Parses the given {@code String} of arguments in the context of the EditContactCommand
-     * and returns an EditContactCommand object for execution.
-     * @throws ParseException if the user input does not conform the expected format
-     */
-    public EditContactCommand parse(String args) throws ParseException {
-        requireNonNull(args);
-        ArgumentTokenizer tokenizer = new ArgumentTokenizer(args, PREFIX_NAME, PREFIX_EMAIL, PREFIX_TELEGRAM);
-        ArgumentMultimap argMultimap = tokenizer.tokenize();
-
-        Index index;
-
-        try {
-            index = ParserUtil.parseIndex(argMultimap.getPreamble());
-        } catch (ParseException pe) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT,
-                    EditContactCommand.MESSAGE_USAGE), pe);
-        }
-
-        EditContactDescriptor editContactDescriptor = new EditContactDescriptor();
-        if (argMultimap.getValue(PREFIX_NAME).isPresent()) {
-            editContactDescriptor.setName(ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get()));
-        }
-
-        if (argMultimap.getValue(PREFIX_EMAIL).isPresent()) {
-            editContactDescriptor.setEmail(ParserUtil.parseEmail(argMultimap.getValue(PREFIX_EMAIL).get()));
-        }
-
-        if (argMultimap.getValue(PREFIX_TELEGRAM).isPresent()) {
-            editContactDescriptor.setTelegram(ParserUtil.parseTelegram(argMultimap.getValue(PREFIX_TELEGRAM).get()));
-        }
-
-        if (!editContactDescriptor.isAnyFieldEdited()) {
-            throw new ParseException(EditContactCommand.MESSAGE_NOT_EDITED);
-        }
-
-        return new EditContactCommand(index, editContactDescriptor);
-    }
-
-}
