+++ src/main/java/seedu/address/logic/parser/ModuleListParser.java
@@ -7,20 +7,23 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import seedu.address.logic.commands.AddAssignmentCommand;
-import seedu.address.logic.commands.contactlistcommands.AddContactCommand;
 import seedu.address.logic.commands.AddModuleCommand;
 import seedu.address.logic.commands.AddZoomLinkCommand;
 import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.contactlistcommands.DeleteContactCommand;
 import seedu.address.logic.commands.DeleteModuleCommand;
-import seedu.address.logic.commands.contactlistcommands.EditContactCommand;
 import seedu.address.logic.commands.EditModuleCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.commands.ViewModuleCommand;
+import seedu.address.logic.commands.contactlistcommands.AddContactCommand;
+import seedu.address.logic.commands.contactlistcommands.DeleteContactCommand;
+import seedu.address.logic.commands.contactlistcommands.EditContactCommand;
+import seedu.address.logic.parser.contactlistparsers.AddContactParser;
+import seedu.address.logic.parser.contactlistparsers.DeleteContactParser;
+import seedu.address.logic.parser.contactlistparsers.EditContactParser;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
