+++ src/main/java/seedu/address/logic/parser/contactlistparsers/ContactListParser.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.contactlistparsers;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
@@ -8,15 +8,13 @@ import java.util.regex.Pattern;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.HelpCommand;
+import seedu.address.logic.commands.contactlistcommands.AddContactCommand;
+import seedu.address.logic.commands.contactlistcommands.DeleteContactCommand;
+import seedu.address.logic.commands.contactlistcommands.EditContactCommand;
 import seedu.address.logic.commands.contactlistcommands.FindContactCommand;
 import seedu.address.logic.commands.contactlistcommands.HelpContactCommand;
 import seedu.address.logic.commands.contactlistcommands.ListContactCommand;
 import seedu.address.logic.commands.contactlistcommands.ViewContactCommand;
-import seedu.address.logic.parser.contactlistparsers.AddContactParser;
-import seedu.address.logic.parser.contactlistparsers.DeleteContactParser;
-import seedu.address.logic.parser.contactlistparsers.EditContactParser;
-import seedu.address.logic.parser.contactlistparsers.FindContactParser;
-import seedu.address.logic.parser.contactlistparsers.ViewContactParser;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 public class ContactListParser {
